Extract delete message tests
